(feat): Organized service Folder
