+++ src/app/body/board/board.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { Sprint, SprintDataId, TeamDataId } from 'src/app/Interface/TeamInterface';
-import { ApplicationSettingsService } from 'src/app/services/application-settings.service';
-import { BackendService } from 'src/app/services/backend.service';
-import { NavbarHandlerService } from 'src/app/services/navbar-handler.service';
+import { ApplicationSettingsService } from 'src/app/services/applicationSettings/application-settings.service';
+import { BackendService } from 'src/app/services/backEnd/backend.service';
+import { NavbarHandlerService } from 'src/app/services/navbar-handler/navbar-handler.service';
 
 @Component({
   selector: 'app-board',
