+++ src/app/body/create-new-session/create-new-session.component.ts
@@ -3,16 +3,16 @@ import { AngularFireFunctions } from '@angular/fire/functions';
 import { NgForm } from '@angular/forms';
 import { Router } from '@angular/router';
 import { User } from 'src/app/Interface/UserInterface';
-import { ValidationService } from '../../services/validation.service';
-import { ToolsService } from '../../services/tools.service';
+import { ValidationService } from '../../services/general/validation.service';
+import { ToolsService } from '../../services/general/tools.service';
 import { Location } from '@angular/common';
-import { NavbarHandlerService } from 'src/app/services/navbar-handler.service';
-import { ErrorHandlerService } from 'src/app/services/error-handler.service';
-import { BackendService } from 'src/app/services/backend.service';
-import { ApplicationSettingsService } from 'src/app/services/application-settings.service';
+import { NavbarHandlerService } from 'src/app/services/navbar-handler/navbar-handler.service';
+import { ErrorHandlerService } from 'src/app/services/error-handler/error-handler.service';
+import { BackendService } from 'src/app/services/backEnd/backend.service';
+import { ApplicationSettingsService } from 'src/app/services/applicationSettings/application-settings.service';
 import { AuthService } from 'src/app/services/auth.service';
 import { Tasks } from 'src/app/Interface/TasksInterface';
-import { CloneTaskService } from 'src/app/services/clone-task.service';
+import { CloneTaskService } from 'src/app/services/cloneTask/clone-task.service';
 
 @Component({
   selector: 'app-create-new-session',
