+++ src/app/body/task-details/task-details.component.ts
@@ -5,13 +5,13 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators'
 import { Tasks } from 'src/app/Interface/TasksInterface';
-import { CloneTaskService } from 'src/app/services/clone-task.service';
+import { CloneTaskService } from 'src/app/services/cloneTask/clone-task.service';
 import { AuthService } from 'src/app/services/auth.service';
-import { ToolsService } from '../../services/tools.service';
+import { ToolsService } from '../../services/general/tools.service';
 import { Location } from '@angular/common';
-import { NavbarHandlerService } from 'src/app/services/navbar-handler.service';
-import { ErrorHandlerService } from 'src/app/services/error-handler.service';
-import { BackendService } from 'src/app/services/backend.service';
+import { NavbarHandlerService } from 'src/app/services/navbar-handler/navbar-handler.service';
+import { ErrorHandlerService } from 'src/app/services/error-handler/error-handler.service';
+import { BackendService } from 'src/app/services/backEnd/backend.service';
 import { Activity, ActivityId } from 'src/app/Interface/ActivityInterface';
 
 @Component({
