--- src/app/services/backend.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { BackendService } from './backend.service';
-
-describe('BackendService', () => {
-  let service: BackendService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(BackendService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
