--- src/app/services/error-handler.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ErrorHandlerService } from './error-handler.service';
-
-describe('ErrorHandlerService', () => {
-  let service: ErrorHandlerService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ErrorHandlerService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
