+++ src/app/user-verification/user-verification.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AngularFireFunctions } from '@angular/fire/functions';
 import { ActivatedRoute, Router } from '@angular/router';
-import { ErrorHandlerService } from 'src/app/services/error-handler.service';
+import { ErrorHandlerService } from 'src/app/services/error-handler/error-handler.service';
 import { AuthService } from '../services/auth.service';
 
 @Component({
